pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "locomp"
version = "0.1.0"
description = "A GPU compute compiler for Apple Metal — write kernels in Python, compile to native Metal shaders"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Locomp Contributors"}
]
keywords = ["gpu", "compiler", "metal", "apple-silicon", "cuda", "compute", "kernel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Compilers",
]
dependencies = [
    "numpy>=1.24.0",
    "pyobjc-framework-Metal>=9.0; sys_platform == 'darwin'",
    "pyobjc-framework-MetalPerformanceShaders>=9.0; sys_platform == 'darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/Zyora-Dev/locomp"
Repository = "https://github.com/Zyora-Dev/locomp"
Documentation = "https://github.com/Zyora-Dev/locomp/wiki"

[tool.setuptools.packages.find]
include = ["locomp*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
